requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
requests==2.32.3
websockets==12.0
livekit==1.0.13
livekit-agents==0.9.1
openai==1.45.0
deepgram-sdk==3.3.0
soundfile==0.12.1
numpy==1.26.4
torch==2.3.1
torchaudio==2.3.1
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.3.1
unstructured==0.14.9
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
python-multipart==0.0.9
pinecone-client[grpc]==5.0.1
